Rebase onto 1.4 and move some preexec code to server/engine
