+++ Unix/server/servercommon.c
@@ -1207,10 +1207,6 @@ void ServerCleanup(int pidfile)
 
 #if defined(USE_ALLOCATOR)
 
-# if defined(USE_PAL_ATEXIT)
-    PAL_AtexitCall();
-# endif
-
     PAL_DumpAllocStats();
 
     if (PAL_GetBlocksAllocated())
